Cattle list management logic detached from view
