+++ app/src/main/java/com/iua/agustinpereyra/view/cattleviews/CattleListFragment.kt
@@ -43,4 +43,8 @@ class CattleListFragment : FilterableCattleRecyclerFragment(){
 
         return view
     }
+
+    interface CattleListFragmentListener : ActionBarModifier {
+        fun navigateToSpecificBovine() : Unit
+    }
 }
